imp: styled-components 폴더로 만들어서 import 하던것 그냥 module 바로 import 하도록 변경
