+++ src/App.tsx
@@ -1,9 +1,12 @@
 import React from 'react';
 
-import styled from '@/styledComponents';
+import styled from 'styled-components';
 
-const Wrapper = styled.div``;
+const Wrapper = styled.div`
+  padding: 1rem 0;
+  text-align: center;
+`;
 
-const App: React.FC = () => <Wrapper>보일러플레이트</Wrapper>;
+const App: React.FC = () => <Wrapper>Photo map :)</Wrapper>;
 
 export default App;
