+++ src/globalStyles.ts
@@ -1,10 +1,8 @@
-import { createGlobalStyle } from '@/styledComponents';
+import { createGlobalStyle } from 'styled-components';
+
+import theme from '@/styles/theme';
 
 const GlobalStyles = createGlobalStyle`
-  /* http://meyerweb.com/eric/tools/css/reset/
-    v4.0 | 20180602
-    License: none (public domain)
-  */
   html, body, div, span, applet, object, iframe,
   h1, h2, h3, h4, h5, h6, p, blockquote, pre,
   a, abbr, acronym, address, big, cite, code,
@@ -25,29 +23,39 @@ const GlobalStyles = createGlobalStyle`
     font: inherit;
     vertical-align: baseline;
   }
-  /* HTML5 display-role reset for older browsers */
+
   article, aside, details, figcaption, figure,
   footer, header, hgroup, main, menu, nav, section {
     display: block;
   }
-  /* HTML5 hidden-attribute fix for newer browsers */
+  
   *[hidden] {
       display: none;
   }
+
+  html, body {
+    font-size: 16px;
+    color: ${theme.color.black};
+  }
+
   body {
     line-height: 1;
   }
+  
   ol, ul {
     list-style: none;
   }
+  
   blockquote, q {
     quotes: none;
   }
+  
   blockquote:before, blockquote:after,
   q:before, q:after {
     content: '';
     content: none;
   }
+  
   table {
     border-collapse: collapse;
     border-spacing: 0;
